Abstract copyElementText into a utility function
